third_party/WebKit/LayoutTests/http/tests/security/dangling-markup/a-ping.html

<!DOCTYPE html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  if (window.testRunner) {
    testRunner.overridePreference("WebKitHyperlinkAuditingEnabled", 1);
    testRunner.dumpPingLoaderCallbacks();
  }
  console.error("This test relies on dumping data via `window.testRunner.dumpPingLoaderCallbacks()`. No URLs containing 'block' should be pinged.");
</script>
</head>
<body>
<script>
  function readableURL(url) {
    return url.replace(/\n/g, "\\n").replace(/\r/g, "\\r").replace(/\t/g, "\\t");
  }

  var should_load = [
    `/resources/square.png?1`,
    `/resources/squ\nare.png?2`,
    `/resources/squ\tare.png?3`,
    `/resources/squ\rare.png?4`,
    `/resources/square.png?img=<&amp;5`,
    `/resources/square.png?img=&lt;&amp;6`,
    `/resources/square.png?img=%3C&amp;7`,
    `/resources/squ\nare.png?img=%3C&amp;8`,
    `/resources/squ\rare.png?img=%3C&amp;9`,
    `/resources/squ\tare.png?img=%3C&amp;10`,
    `/resources/square.png?img=&#10;&amp;11`,
    `/resources/squ\nare.png?img=&#10;&amp;12`,
    `/resources/squ\rare.png?img=&#10;&amp;13`,
    `/resources/squ\tare.png?img=&#10;&amp;14`,
  ];
  should_load.forEach(url => test(t => {
    var a = document.createElement('a');
    a.ping = (new URL(url, document.baseURI).href);
    a.href = '#' + url;
    document.body.appendChild(a);
    a.click();
  }, "Ping: " + readableURL(url)));

  var should_block = [
    `/resources/squ\nare.png?block=<`,
    `/resources/squ\rare.png?block=<`,
    `/resources/squ\tare.png?block=<`,
    `/resources/square.png?<\n=block`,
    `/resources/square.png?<\r=block`,
    `/resources/square.png?<\t=block`,
  ];
  should_block.forEach(url => test(t => {
    var a = document.createElement('a');
    a.ping = document.origin + url;
    a.href = '#' + url;
    document.body.appendChild(a);
    a.click();
  }, "Prefetch: " + readableURL(url)));
</script>
