third_party/WebKit/LayoutTests/http/tests/security/dangling-markup/a-ping-expected.txt

CONSOLE ERROR: line 10: This test relies on dumping data via `window.testRunner.dumpPingLoaderCallbacks()`. No URLs containing 'block' should be pinged.
PingLoader dispatched to 'http://127.0.0.1:8000/resources/square.png?1'.
PingLoader dispatched to 'http://127.0.0.1:8000/resources/square.png?2'.
PingLoader dispatched to 'http://127.0.0.1:8000/resources/square.png?3'.
PingLoader dispatched to 'http://127.0.0.1:8000/resources/square.png?4'.
PingLoader dispatched to 'http://127.0.0.1:8000/resources/square.png?img=%3C&amp;5'.
PingLoader dispatched to 'http://127.0.0.1:8000/resources/square.png?img=&lt;&amp;6'.
PingLoader dispatched to 'http://127.0.0.1:8000/resources/square.png?img=%3C&amp;7'.
PingLoader dispatched to 'http://127.0.0.1:8000/resources/square.png?img=%3C&amp;8'.
PingLoader dispatched to 'http://127.0.0.1:8000/resources/square.png?img=%3C&amp;9'.
PingLoader dispatched to 'http://127.0.0.1:8000/resources/square.png?img=%3C&amp;10'.
PingLoader dispatched to 'http://127.0.0.1:8000/resources/square.png?img=&#10;&amp;11'.
PingLoader dispatched to 'http://127.0.0.1:8000/resources/square.png?img=&#10;&amp;12'.
PingLoader dispatched to 'http://127.0.0.1:8000/resources/square.png?img=&#10;&amp;13'.
PingLoader dispatched to 'http://127.0.0.1:8000/resources/square.png?img=&#10;&amp;14'.
CONSOLE WARNING: line 56: Resource requests whose URLs contained both removed whitespace (`\n`, `\r`, `\t`) characters and less-than characters (`<`) are blocked. Please remove newlines and encode less-than characters from places like element attribute values in order to load these resources. See https://www.chromestatus.com/feature/5735596811091968 for more details.
This is a testharness.js-based test.
PASS Ping: /resources/square.png?1
PASS Ping: /resources/squ\nare.png?2
PASS Ping: /resources/squ\tare.png?3
PASS Ping: /resources/squ\rare.png?4
PASS Ping: /resources/square.png?img=<&amp;5
PASS Ping: /resources/square.png?img=&lt;&amp;6
PASS Ping: /resources/square.png?img=%3C&amp;7
PASS Ping: /resources/squ\nare.png?img=%3C&amp;8
PASS Ping: /resources/squ\rare.png?img=%3C&amp;9
PASS Ping: /resources/squ\tare.png?img=%3C&amp;10
PASS Ping: /resources/square.png?img=&#10;&amp;11
PASS Ping: /resources/squ\nare.png?img=&#10;&amp;12
PASS Ping: /resources/squ\rare.png?img=&#10;&amp;13
PASS Ping: /resources/squ\tare.png?img=&#10;&amp;14
PASS Prefetch: /resources/squ\nare.png?block=<
PASS Prefetch: /resources/squ\rare.png?block=<
PASS Prefetch: /resources/squ\tare.png?block=<
PASS Prefetch: /resources/square.png?<\n=block
PASS Prefetch: /resources/square.png?<\r=block
PASS Prefetch: /resources/square.png?<\t=block
Harness: the test ran to completion.
